move and split Parameters
